Home Work 5
The analysis focused on assessing the effectiveness of advertising campaigns on Google and Facebook platforms. Data on expenses, impressions, clicks, and conversion volumes were collected for different dates and campaigns. The collected data underwent processing, resulting in key metrics such as expenses, impressions, clicks, conversion volume, CTR (click-through rate), CPC (cost per click), and CPM (cost per mille). Additionally, ROMI (Return on Marketing Investment) was calculated, representing the ratio of revenue from the advertising campaign to the expenses incurred.

As part of the bonus task, a code snippet was discovered via a search engine for creating a function to decode Cyrillic utm tags. The code was adapted to fit my database and a function was created accordingly. This function was then applied to the main task.